# src/main/resources/swagger/MS-Address-API.yaml
openapi: 3.0.3
info:
  title: MS Address API
  description: |
    This API provides operations related to client addresses. It follows the OpenAPI 3.0 specification.
  contact:
    email: dowglasmaia@live.com
  version: 1.0.0
tags:
  - name: Address
    description: Operations related to client addresses
paths:
  /addresses/findByZipCode:
    get:
      tags:
        - Address
      summary: Find Address by Zipcode
      description: Retrieve the address of a client by their ID.
      operationId: findByZipCode
      parameters:
        - name: zipCode
          in: query
          description: ZipCodeID of the Address
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddressResponse'
        '404':
          description: Address not found
        '500':
          description: Internal server error
        '503':
          description: Service unavailable
  /addresses:
    post:
      tags:
        - Address
      summary: Create Address
      description: Create a new address.
      operationId: createAddress
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddressRequest'
      responses:
        '201':
          description: Address created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddressIdResponse'
        '400':
          description: Invalid request body or parameters
        '500':
          description: Internal server error
        '503':
          description: Service unavailable
components:
  schemas:
    AddressRequest:
      type: object
      properties:
        street:
          type: string
          description: Street address
          example: 437 Lytton Ave
        number:
          type: string
          description: Street number
          example: 437
        city:
          type: string
          description: City
          example: Palo Alto
        state:
          type: string
          description: State abbreviation
          enum: [AM, MA, SC, PR, PI]  # Definindo o campo state como um enum com opções específicas
          example: AM
        zip:
          type: string
          description: ZIP code
          example: '94301'
    AddressResponse:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier for the address
        street:
          type: string
          description: Street address
          example: 437 Lytton Ave
        number:
          type: string
          description: Street number
          example: 437
        city:
          type: string
          description: City
          example: Palo Alto
        state:
          type: string
          description: State abbreviation
          example: CA
        zip:
          type: string
          description: ZIP code
          example: '94301'

    AddressIdResponse:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier for the address
          example: abc123